extract less functions for reuse
- could be the base for a future extension or design-time UI
